feature: handle MAV_CMD_DO_SET_CAM_TRIGG_DIST & MAV_CMD_DO_DIGICAM_CONTROL in Missions by calling mobile camera.

+++ andruavProtocol/src/main/java/com/andruav/controlBoard/shared/missions/MissionDayra.java
@@ -6,7 +6,7 @@ package com.andruav.controlBoard.shared.missions;
 public class MissionDayra extends MissionBase {
 
 
-    public final static byte TYPE_DAYRA = 5;
+    public final static int TYPE_DAYRA = 5;
 
     public double Altitude;
     public double Latitude;
